# examples/petstore/lib/testdata/doc/openapi.golden.json
{
	"components": {
		"schemas": {
			"HTTPError": {
				"description": "HTTPError schema",
				"properties": {
					"detail": {
						"description": "Human readable error message",
						"nullable": true,
						"type": "string"
					},
					"errors": {
						"items": {
							"properties": {
								"more": {
									"additionalProperties": {},
									"type": "object"
								},
								"name": {
									"type": "string"
								},
								"reason": {
									"type": "string"
								}
							},
							"type": "object"
						},
						"nullable": true,
						"type": "array"
					},
					"instance": {
						"nullable": true,
						"type": "string"
					},
					"status": {
						"description": "HTTP status code",
						"example": 403,
						"nullable": true,
						"type": "integer"
					},
					"title": {
						"description": "Short title of the error",
						"nullable": true,
						"type": "string"
					},
					"type": {
						"description": "URL of the error type. Can be used to lookup the error in a documentation",
						"nullable": true,
						"type": "string"
					}
				},
				"type": "object"
			},
			"NoContent": {
				"description": "NoContent schema"
			},
			"Pets": {
				"description": "Pets schema",
				"properties": {
					"age": {
						"description": "Age of the pet, in years",
						"example": 2,
						"type": "integer"
					},
					"id": {
						"example": "pet-123456",
						"type": "string"
					},
					"is_adopted": {
						"description": "Is the pet adopted",
						"type": "boolean"
					},
					"name": {
						"example": "Napoleon",
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object"
			},
			"PetsCreate": {
				"description": "PetsCreate schema",
				"properties": {
					"age": {
						"description": "Age of the pet, in years",
						"example": 2,
						"maximum": 100,
						"minimum": 0,
						"type": "integer"
					},
					"is_adopted": {
						"description": "Is the pet adopted",
						"type": "boolean"
					},
					"name": {
						"example": "Napoleon",
						"maxLength": 100,
						"minLength": 1,
						"type": "string"
					}
				},
				"required": [
					"name"
				],
				"type": "object"
			},
			"PetsError": {
				"description": "PetsError schema",
				"properties": {
					"message": {
						"type": "string"
					}
				},
				"type": "object"
			},
			"PetsUpdate": {
				"description": "PetsUpdate schema",
				"properties": {
					"age": {
						"example": 2,
						"maximum": 100,
						"nullable": true,
						"type": "integer"
					},
					"is_adopted": {
						"description": "Is the pet adopted",
						"nullable": true,
						"type": "boolean"
					},
					"name": {
						"description": "Name of the pet",
						"example": "Napoleon",
						"maxLength": 100,
						"minLength": 1,
						"nullable": true,
						"type": "string"
					}
				},
				"type": "object"
			},
			"unknown-interface": {
				"description": "unknown-interface schema"
			}
		}
	},
	"info": {
		"description": "\nThis is the autogenerated OpenAPI documentation for your [Fuego](https://github.com/go-fuego/fuego) API.\n\nBelow is a Fuego Cheatsheet to help you get started. Don't hesitate to check the [Fuego documentation](https://go-fuego.github.io/fuego) for more details.\n\nHappy coding! 🔥\n\n## Usage\n\n### Route registration\n\n```go\nfunc main() {\n\t// Create a new server\n\ts := fuego.NewServer()\n\n\t// Register some routes\n\tfuego.Post(s, \"/hello\", myController)\n\tfuego.Get(s, \"/myPath\", otherController)\n\tfuego.Put(s, \"/hello\", thirdController)\n\n\tadminRoutes := fuego.Group(s, \"/admin\")\n\tfuego.Use(adminRoutes, myMiddleware) // This middleware (for authentication, etc...) will be available for routes starting by /admin/*, \n\tfuego.Get(adminRoutes, \"/hello\", groupController) // This route will be available at /admin/hello\n\n\t// Start the server\n\ts.Start()\n}\n```\n\n### Basic controller\n\n```go\ntype MyBody struct {\n\tName string `json:\"name\" validate:\"required,max=30\"`\n}\n\ntype MyResponse struct {\n\tAnswer string `json:\"answer\"`\n}\n\nfunc hello(ctx fuego.ContextWithBody[MyBody]) (*MyResponse, error) {\n\tbody, err := ctx.Body()\n\tif err != nil {\n\t\treturn nil, err\n\t}\n\n\treturn \u0026MyResponse{Answer: \"Hello \" + body.Name}, nil\n}\n```\n\n### Add openAPI information to the route\n\n```go\nimport (\n\t\"github.com/go-fuego/fuego\"\n\t\"github.com/go-fuego/fuego/option\"\n\t\"github.com/go-fuego/fuego/param\"\n)\n\nfunc main() {\n\ts := fuego.NewServer()\n\n\t// Custom OpenAPI options\n\tfuego.Post(s, \"/\", myController\n\t\toption.Description(\"This route does something...\"),\n\t\toption.Summary(\"This is my summary\"),\n\t\toption.Tags(\"MyTag\"), // A tag is set by default according to the return type (can be deactivated)\n\t\toption.Deprecated(), // Marks the route as deprecated in the OpenAPI spec\n\n\t\toption.Query(\"name\", \"Declares a query parameter with default value\", param.Default(\"Carmack\")),\n\t\toption.Header(\"Authorization\", \"Bearer token\", param.Required()),\n\t\toptionPagination,\n\t\toptionCustomBehavior,\n\t)\n\n\ts.Run()\n}\n\nvar optionPagination = option.Group(\n\toption.QueryInt(\"page\", \"Page number\", param.Default(1), param.Example(\"1st page\", 1), param.Example(\"42nd page\", 42)),\n\toption.QueryInt(\"perPage\", \"Number of items per page\"),\n)\n\nvar optionCustomBehavior = func(r *fuego.BaseRoute) {\n\tr.XXX = \"YYY\"\n}\n```\n\nThen, in the controller\n\n```go\ntype MyResponse struct {\n\tAnswer string `json:\"answer\"`\n}\n\nfunc getAllPets(ctx fuego.ContextNoBody) (*MyResponse, error) {\n\tname := ctx.QueryParam(\"name\")\n\tperPage, _ := ctx.QueryParamIntErr(\"per_page\")\n\n\treturn \u0026MyResponse{Answer: \"Hello \" + name}, nil\n}\n```\n",
		"title": "OpenAPI",
		"version": "0.0.1"
	},
	"openapi": "3.1.0",
	"paths": {
		"/pets/": {
			"get": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.filterPets`\n\n---\n\nFilter pets",
				"operationId": "GET_/pets/",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Number of items per page",
						"in": "query",
						"name": "per_page",
						"required": true,
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Page number",
						"examples": {
							"100th page": {
								"value": 100
							},
							"1st page": {
								"value": 1
							},
							"42nd page": {
								"value": 42
							}
						},
						"in": "query",
						"name": "page",
						"schema": {
							"default": 1,
							"type": "integer"
						}
					},
					{
						"description": "Filter by name",
						"examples": {
							"cat name": {
								"value": "felix"
							}
						},
						"in": "query",
						"name": "name",
						"schema": {
							"nullable": true,
							"type": "string"
						}
					},
					{
						"description": "Only get pets younger than given age in years",
						"in": "query",
						"name": "younger_than",
						"schema": {
							"default": 3,
							"type": "integer"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pets"
									},
									"type": "array"
								}
							},
							"application/xml": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pets"
									},
									"type": "array"
								}
							}
						},
						"description": "OK",
						"headers": {
							"Content-Range": {
								"examples": {
									"42 pets": {
										"value": "0-10/42"
									}
								},
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"206": {
						"description": "OK",
						"headers": {
							"Content-Range": {
								"examples": {
									"42 pets": {
										"value": "0-10/42"
									}
								},
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "filter pets",
				"tags": [
					"pets"
				]
			},
			"post": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.postPets`\n\n---\n\n",
				"operationId": "POST_/pets/",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/PetsCreate"
							}
						}
					},
					"description": "Request body for models.PetsCreate",
					"required": true
				},
				"responses": {
					"201": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							}
						},
						"description": "Created"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"409": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PetsError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/PetsError"
								}
							}
						},
						"description": "Conflict: Pet with the same name already exists"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "post pets",
				"tags": [
					"pets"
				]
			}
		},
		"/pets/all": {
			"get": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.getAllPets`\n\n---\n\nGet all pets",
				"operationId": "GET_/pets/all",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Number of items per page",
						"in": "query",
						"name": "per_page",
						"required": true,
						"schema": {
							"type": "integer"
						}
					},
					{
						"description": "Page number",
						"examples": {
							"100th page": {
								"value": 100
							},
							"1st page": {
								"value": 1
							},
							"42nd page": {
								"value": 42
							}
						},
						"in": "query",
						"name": "page",
						"schema": {
							"default": 1,
							"type": "integer"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pets"
									},
									"type": "array"
								}
							},
							"application/xml": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pets"
									},
									"type": "array"
								}
							}
						},
						"description": "OK",
						"headers": {
							"Content-Range": {
								"examples": {
									"42 pets": {
										"value": "0-10/42"
									}
								},
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"206": {
						"description": "OK",
						"headers": {
							"Content-Range": {
								"examples": {
									"42 pets": {
										"value": "0-10/42"
									}
								},
								"schema": {
									"type": "string"
								}
							}
						}
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "get all pets",
				"tags": [
					"pets",
					"my-tag"
				]
			}
		},
		"/pets/by-age": {
			"get": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.getAllPetsByAge`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego/examples/petstore/controllers.dummyMiddleware`\n\n---\n\nReturns an array of pets grouped by age",
				"operationId": "GET_/pets/by-age",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"items": {
											"$ref": "#/components/schemas/Pets"
										},
										"type": "array"
									},
									"type": "array"
								}
							},
							"application/xml": {
								"schema": {
									"items": {
										"items": {
											"$ref": "#/components/schemas/Pets"
										},
										"type": "array"
									},
									"type": "array"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "get all pets by age",
				"tags": [
					"pets"
				]
			}
		},
		"/pets/by-name/{name...}": {
			"get": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.getPetByName`\n\n---\n\n",
				"operationId": "GET_/pets/by-name/:name...",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": " (might contain slashes)",
						"in": "path",
						"name": "name...",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "get pet by name",
				"tags": [
					"pets"
				]
			}
		},
		"/pets/std/all": {
			"get": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.Routes.func1`\n\n---\n\n",
				"operationId": "GET_/pets/std/all",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Pets"
									},
									"type": "array"
								}
							}
						},
						"description": "all the pets"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "func1",
				"tags": [
					"pets",
					"std"
				]
			}
		},
		"/pets/{id}": {
			"delete": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.deletePets`\n\n---\n\n",
				"operationId": "DELETE_/pets/:id",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/unknown-interface"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/unknown-interface"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "delete pets",
				"tags": [
					"pets"
				]
			},
			"get": {
				"description": "Replace description with this sentence.",
				"operationId": "getPet",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Pet ID",
						"examples": {
							"example": {
								"value": "123"
							}
						},
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "get pets",
				"tags": [
					"pets"
				]
			},
			"put": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.putPets`\n\n---\n\n",
				"operationId": "PUT_/pets/:id",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"*/*": {
							"schema": {
								"$ref": "#/components/schemas/PetsUpdate"
							}
						}
					},
					"description": "Request body for models.PetsUpdate",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "put pets",
				"tags": [
					"pets"
				]
			}
		},
		"/pets/{id}/json": {
			"put": {
				"description": "#### Controller: \n\n`github.com/go-fuego/fuego/examples/petstore/controllers.PetsResources.putPets`\n\n---\n\n",
				"operationId": "PUT_/pets/:id/json",
				"parameters": [
					{
						"description": "header description",
						"in": "header",
						"name": "X-Header",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "Accept",
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PetsUpdate"
							}
						}
					},
					"description": "Request body for models.PetsUpdate",
					"required": true
				},
				"responses": {
					"200": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/Pets"
								}
							}
						},
						"description": "OK"
					},
					"204": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/NoContent"
								}
							}
						},
						"description": "No Content"
					},
					"400": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Bad Request _(validation or deserialization error)_"
					},
					"500": {
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							},
							"application/xml": {
								"schema": {
									"$ref": "#/components/schemas/HTTPError"
								}
							}
						},
						"description": "Internal Server Error _(panics)_"
					},
					"default": {
						"description": ""
					}
				},
				"summary": "Update a pet with JSON-only body",
				"tags": [
					"pets"
				]
			}
		}
	},
	"servers": [
		{
			"description": "local server",
			"url": "http://localhost:9999"
		}
	],
	"tags": [
		{
			"name": "my-tag"
		},
		{
			"name": "pets"
		},
		{
			"name": "std"
		}
	]
}